Refactor the home page
