+++ client/src/api/announces.ts
@@ -3,8 +3,8 @@ import type { Book } from "./types";
 
 export async function announcesList(): Promise<Book[]> {
     try{
-      const { data } = await axios.get("/api/book/announce/");
-      console.log("/api/book/announce/", data);
+      const { data } = await axios.get("/api/book/announcement/");
+      console.log("/api/book/announcement/", data);
       return data;
     }catch (error) {
       console.error("Ошибка при поиске новинок", error);
